Compile slurm for the os of the submitter
So that it can submit jobs to a cluster that uses a different OS.

Automate the configuration of the cluster manager and submitters in RES.

Fix errors in configuration scripts.

Add a parameter to the ParallelCluster custom resource that has a hash
of all of the config file contents so that the resource gets updated
whenever anything changes.
This current issue with this implementation is that the ParallelCluster
still doesn't change so even though the custom resource is updated,
the cluster doesn't because it's configuration is unchanged.

Fix a bug in configuring licenses.
After initial configuration the playbook was failing on subsequent updates.

Support custom munge key which must be specified in secretsmanager

Try to allow RES /home to be mounted, but currently causes a validation error.

Update EDA instance types to the latest and so that all memory sizes get built.

Restrict lambda function IAM roles and cleanup cfn_nag errors and warnings.

Resolves #177
Resolves #178
Resolves #180
Resolves #181

+++ source/resources/parallel-cluster/config/bin/configure-eda.sh
@@ -1,10 +1,25 @@
-#!/bin/bash -x
+#!/bin/bash -ex
+# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# SPDX-License-Identifier: MIT-0
+
+# This script calls an ansible playbook that installs packages typically used by EDA tools.
+# It has 2 different use cases:
+#   * To build ParallelCluster AMIs
+#   * To install packages on VDIs or other login nodes using a ParallellCluster.
+# The location of the config directory is different for those 2 use cases.
+# For an AMI build, the config directory and scripts will not exist and must be downloaded from S3.
+# For a login node, the playbooks and scripts will already exist.
 
 script=$0
+script_name=$(basename $script)
 
-AWS_DEFAULT_REGION={{Region}}
+# Jinja2 template variables
+assets_bucket={{assets_bucket}}
+assets_base_key={{assets_base_key}}
+export AWS_DEFAULT_REGION={{Region}}
 ClusterName={{ClusterName}}
 ErrorSnsTopicArn={{ErrorSnsTopicArn}}
+playbooks_s3_url={{playbooks_s3_url}}
 
 # Notify user of errors
 function on_exit {
@@ -24,45 +39,47 @@ trap on_exit EXIT
 # Redirect all IO to /var/log/messages and then echo to stderr
 exec 1> >(logger -s -t PCImageBuilderEDA) 2>&1
 
-assets_bucket={{assets_bucket}}
-assets_base_key={{assets_base_key}}
-playbooks_s3_url={{playbooks_s3_url}}
-
-ansible_head_node_vars_yml_s3_url="s3://$assets_bucket/$assets_base_key/config/ansible/ansible_head_node_vars.yml"
-ansible_compute_node_vars_yml_s3_url="s3://$assets_bucket/$assets_base_key/config/ansible/ansible_compute_node_vars.yml"
-ansible_submitter_vars_yml_s3_url="s3://$assets_bucket/$assets_base_key/config/ansible/ansible_submitter_vars.yml"
-
-config_dir=/opt/slurm/config
-config_bin_dir=$config_dir/bin
-
-mkdir -p $config_bin_dir
-
 # Install ansible
 if ! yum list installed ansible &> /dev/null; then
     yum install -y ansible || amazon-linux-extras install -y ansible2
 fi
 
-# Download ansible playbooks
-
+submitter_config_dir=/opt/slurm/${ClusterName}/config
+if [ -e $submitter_config_dir ]; then
+    config_dir=$submitter_config_dir
+else
+    config_dir=/opt/slurm/config
+fi
+config_bin_dir=$config_dir/bin
 ANSIBLE_PATH=$config_dir/ansible
 PLAYBOOKS_PATH=$ANSIBLE_PATH/playbooks
 PLAYBOOKS_ZIP_PATH=$ANSIBLE_PATH/playbooks.zip
-aws s3 cp $playbooks_s3_url ${PLAYBOOKS_ZIP_PATH}.new
-if ! [ -e $PLAYBOOKS_ZIP_PATH ] || ! diff -q $PLAYBOOKS_ZIP_PATH ${PLAYBOOKS_ZIP_PATH}.new; then
-    mv $PLAYBOOKS_ZIP_PATH.new $PLAYBOOKS_ZIP_PATH
-    rm -rf $PLAYBOOKS_PATH
-    mkdir -p $PLAYBOOKS_PATH
-    pushd $PLAYBOOKS_PATH
-    unzip $PLAYBOOKS_ZIP_PATH
-    chmod -R 0700 $ANSIBLE_PATH
-    popd
+
+if ! [ -e $submitter_config_dir ]; then
+    mkdir -p $config_bin_dir
+
+    ansible_head_node_vars_yml_s3_url="s3://$assets_bucket/$assets_base_key/config/ansible/ansible_head_node_vars.yml"
+    ansible_compute_node_vars_yml_s3_url="s3://$assets_bucket/$assets_base_key/config/ansible/ansible_compute_node_vars.yml"
+    ansible_submitter_vars_yml_s3_url="s3://$assets_bucket/$assets_base_key/config/ansible/ansible_submitter_vars.yml"
+
+    # Download ansible playbooks
+    aws s3 cp $playbooks_s3_url ${PLAYBOOKS_ZIP_PATH}.new
+    if ! [ -e $PLAYBOOKS_ZIP_PATH ] || ! diff -q $PLAYBOOKS_ZIP_PATH ${PLAYBOOKS_ZIP_PATH}.new; then
+        mv $PLAYBOOKS_ZIP_PATH.new $PLAYBOOKS_ZIP_PATH
+        rm -rf $PLAYBOOKS_PATH
+        mkdir -p $PLAYBOOKS_PATH
+        pushd $PLAYBOOKS_PATH
+        unzip $PLAYBOOKS_ZIP_PATH
+        chmod -R 0700 $ANSIBLE_PATH
+        popd
+    fi
+
+    aws s3 cp $ansible_head_node_vars_yml_s3_url /opt/slurm/config/ansible/ansible_head_node_vars.yml
+
+    aws s3 cp $ansible_compute_node_vars_yml_s3_url /opt/slurm/config/ansible/ansible_compute_node_vars.yml
+
+    aws s3 cp $ansible_submitter_vars_yml_s3_url /opt/slurm/config/ansible/ansible_submitter_vars.yml
 fi
-
-aws s3 cp $ansible_head_node_vars_yml_s3_url /opt/slurm/config/ansible/ansible_head_node_vars.yml
-
-aws s3 cp $ansible_compute_node_vars_yml_s3_url /opt/slurm/config/ansible/ansible_compute_node_vars.yml
-
-aws s3 cp $ansible_submitter_vars_yml_s3_url /opt/slurm/config/ansible/ansible_submitter_vars.yml
 
 pushd $PLAYBOOKS_PATH
